+++ CHANGELOG.md
@@ -1,6 +1,28 @@
 # InfinityOne Desktop Client Changelog
 
+## 2.0.0-alpha3 (2020-08-28)
+
+### Bug Fixes
+
+* Fix fetch config settings in main process
+* Fix Start Minimized setting
+* Fix Beta Updates setting
+* Fix vuex migration sync to main for exec mutation type
+
+### Fetures
+
+* Change title to InfinityOne Desktop
+* Add (dev) do title when running in dev
+
+### Known Issues
+
+* Windows and Linux not yet supported
+* Video window may lockup
+* Spell checker not working
+* localization not supported
+
 ## 2.0.0-alpha2 (2020-08-28)
+
 
 ### Bug Fixes